requirements.txt
google-adk
python-dotenv
ragflow-sdk
loguru
litellm
python-docx 
openpyxl 
python-pptx